+++ src/Helpers/FileHelper.php
@@ -2,19 +2,19 @@
 
 namespace seregazhuk\PinterestBot\Helpers;
 
-use seregazhuk\PinterestBot\Exceptions\InvalidRequestException;
+use seregazhuk\PinterestBot\Exceptions\InvalidRequest;
 
 class FileHelper
 {
     /**
      * @param string $filePath
      * @return mixed
-     * @throws InvalidRequestException
+     * @throws InvalidRequest
      */
     public static function getMimeType($filePath)
     {
         if (!file_exists($filePath)) {
-            throw new InvalidRequestException("$filePath: failed to open file.");
+            throw new InvalidRequest("$filePath: failed to open file.");
         }
         
         $fileInfo = finfo_open(FILEINFO_MIME_TYPE);