+++ tests/Api/InterestsTest.php
@@ -2,7 +2,7 @@
 
 namespace seregazhuk\tests\Api;
 
-use seregazhuk\PinterestBot\Helpers\UrlHelper;
+use seregazhuk\PinterestBot\Helpers\UrlBuilder;
 use seregazhuk\tests\Helpers\FollowResponseHelper;
 use seregazhuk\PinterestBot\Api\Providers\Interests;
 
@@ -27,10 +27,10 @@ class InterestsTest extends ProviderTest
     public function it_should_follow_interest()
     {
         $interestId = 1111;
-        $this->setFollowSuccessResponse($interestId, UrlHelper::RESOURCE_FOLLOW_INTEREST);
+        $this->setFollowSuccessResponse($interestId, UrlBuilder::RESOURCE_FOLLOW_INTEREST);
         $this->assertTrue($this->provider->follow($interestId));
 
-        $this->setFollowErrorResponse($interestId, UrlHelper::RESOURCE_FOLLOW_INTEREST);
+        $this->setFollowErrorResponse($interestId, UrlBuilder::RESOURCE_FOLLOW_INTEREST);
         $this->assertFalse($this->provider->follow($interestId));
     }
 
@@ -38,10 +38,10 @@ class InterestsTest extends ProviderTest
     public function it_should_unfollow_interest()
     {
         $interestId = 1111;
-        $this->setFollowSuccessResponse($interestId, UrlHelper::RESOURCE_UNFOLLOW_INTEREST);
+        $this->setFollowSuccessResponse($interestId, UrlBuilder::RESOURCE_UNFOLLOW_INTEREST);
         $this->assertTrue($this->provider->unFollow($interestId));
 
-        $this->setFollowErrorResponse($interestId, UrlHelper::RESOURCE_UNFOLLOW_INTEREST);
+        $this->setFollowErrorResponse($interestId, UrlBuilder::RESOURCE_UNFOLLOW_INTEREST);
         $this->assertFalse($this->provider->unFollow($interestId));
     }